refactor: remane testlib.h to funtestlib.h
feat: change Makefile
