+++ Makefile
@@ -1,46 +1,66 @@
+# Library name
+LIB_NAME = libfuntest.a
+
 # Compiler and flags for compilation
 CC = cc
 CFLAGS = -Wall -Werror -Wextra
 MDIR = mkdir -p
-RM = rm
+RM = rm -f
+AR = ar rcs
 
-## Test library directories
-# uncoment TEST_LIB_ROOT value once in the app 
-TEST_LIB_ROOT = # fun_testlib/
-TEST_DIR_SRC = $(TEST_LIB_ROOT)src
-TEST_BIN_DIR = $(TEST_LIB_ROOT)bin
+# Directories
+INCLUDE = include
+LIB = lib
+SRC = src
 
+## Library source and objects
+SRCS = $(wildcard $(SRC)/*.c)
+OBJS = $(SRCS:.c=.o)
+
+# Tasks
+all: $(LIB_NAME)
+
+# create objects and packaging
+$(LIB_NAME): $(OBJS)
+	$(AR) $(LIB)/$(LIB_NAME) $(OBJS)
+
+%.o: %.c
+	$(CC) $(CFLAGS) -Iinclude -c $< -o $@
+
+# clean directories
+clean:
+	$(RM) $(SRC)/*.o
+
+fclean: clean
+	$(RM) $(LIB)/$(LIB_NAME)
+
+# relink
+re: fclean all
+
+# TEST Section
 ## Name of the test main executable
-TEST_BIN = fun_test_lib.o
+TEST_BIN = test.out
 
-## Library Test source
-TEST_SRC = $(wildcard $(TEST_DIR_SRC)/*.c)
+## Library Test source and objects
+TEST_SRCS = $(wildcard src-tests/*.c)
+TEST_OBJS = $(TEST_SRCS:.c=.o)
 
-## Test source apps
-TEST_DIR_APP = $(TEST_LIB_ROOT)src-tests
-TEST_SRC_APP = $(wildcard $(TEST_DIR_APP)/*.c)
-TESTED_LIBS = # ADD APPS LIBRARIES HERE
-
-dirs:
-	$(MDIR) $(TEST_BIN_DIR)
-
-# Test: generate test binary and launch it
-test: dirs
-	$(CC) $(CFLAGS) -I$(TEST_LIB_ROOT)include -lbsd $(TEST_SRC) $(TEST_SRC_APP) $(TESTED_LIBS) -o $(TEST_BIN_DIR)/$(TEST_BIN)
-	./$(TEST_BIN_DIR)/$(TEST_BIN)
+# Test: generate test binary
+test: $(TEST_OBJS)
+	$(CC) $(CFLAGS) -I$(INCLUDE) $(TEST_SRCS) -L$(LIB) -l:$(LIB_NAME) -o $(TEST_BIN)
+	./$(TEST_BIN)
 
 # clean test directories
-clean:
-	$(RM) -fr $(TEST_BIN_DIR)
+tclean:
+	$(RM) $(TEST_OBJS) $(TEST_BIN)
 
-# detect memory leaks
-sane:
-	$(CC) $(CFLAGS) -I$(TEST_LIB_ROOT)include -lbsd $(TEST_SRC) $(TEST_SRC_APP) $(TESTED_LIBS) -o $(TEST_BIN_DIR)/$(TEST_BIN) -fsanitize=address -g
-	./$(TEST_BIN_DIR)/$(TEST_BIN)
+# Memory leaks detection
+sane: all
+	$(CC) $(CFLAGS) -I$(INCLUDE) $(TEST_SRCS) -L$(LIB) -l:$(LIB_NAME) -o $(TEST_BIN) -fsanitize=address -g
+	./$(TEST_BIN)
 
-val:
-	$(CC) $(CFLAGS) -I$(TEST_LIB_ROOT)include -lbsd $(TEST_SRC) $(TEST_SRC_APP) $(TESTED_LIBS) -o $(TEST_BIN_DIR)/$(TEST_BIN) -fsanitize=address -g
-	valgrind --leak-check=full ./$(TEST_BIN_DIR)/$(TEST_BIN)
+val: all
+	$(CC) $(CFLAGS) -I$(INCLUDE) $(TEST_SRCS) -L$(LIB) -l:$(LIB_NAME) -o $(TEST_BIN)
+	valgrind --leak-check=full ./$(TEST_BIN)
 
-
-.PHONY: test clean dirs
+.PHONY: all clean fclean re sane val
