+++ src-tests/test.h
@@ -5,7 +5,7 @@
 #ifndef TEST_H
 # define TEST_H
 
-# include "testlib.h"
+# include "funtestlib.h"
 
 // add here funcions prototipe
 int	test_ft_dummy(int active);
